consolidate helper functions into single file
